# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodePath Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /progress/course/{courseId}:
    get:
      operationId: getCourseProgress
      tags: [progress]
      summary: Get progress for a specific course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseProgress"

  # GAMIFICATION
...
components:
  schemas:
...
    CourseProgress:
      type: object
      properties:
        courseId:
          type: integer
        courseTitle:
          type: string
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        percentComplete:
          type: integer
        xpEarned:
          type: integer
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - courseId
        - courseTitle
        - completedLessons
        - totalLessons
        - percentComplete
        - xpEarned

    # Gamification
...